+++ common-ts/src/clients/dataApiWsClient.ts
@@ -1,7 +1,7 @@
 import { CandleResolution } from '@drift-labs/sdk';
 import WS from 'isomorphic-ws';
 import { EnvironmentConstants } from '../EnvironmentConstants';
-import { MarketSymbol, UIEnv } from '../types';
+import { JsonCandle, JsonTrade, MarketSymbol, UIEnv } from '../types';
 import { Observable, Subject } from 'rxjs';
 
 const getBaseDataApiUrl = (env: UIEnv) => {
@@ -9,62 +9,6 @@ const getBaseDataApiUrl = (env: UIEnv) => {
 		env.isStaging ? 'staging' : env.isDevnet ? 'dev' : 'mainnet';
 	const dataApiUrl = EnvironmentConstants.dataServerUrl[constantEnv];
 	return dataApiUrl.replace('https://', '');
-};
-
-// Type for the candles returned by the data API
-export type JsonCandle = {
-	ts: number;
-	fillOpen: number;
-	fillHigh: number;
-	fillClose: number;
-	fillLow: number;
-	oracleOpen: number;
-	oracleHigh: number;
-	oracleClose: number;
-	oracleLow: number;
-	quoteVolume: number;
-	baseVolume: number;
-};
-
-export type JsonTrade = {
-	action: string;
-	actionExplanation: string;
-	baseAssetAmountFilled: number;
-	bitFlags: number;
-	createdAt: number;
-	entity: string;
-	fillRecordId: string;
-	filler: string;
-	fillerReward: number;
-	maker: string;
-	makerFee: number;
-	makerOrderBaseAssetAmount: number;
-	makerOrderCumulativeBaseAssetAmountFilled: number;
-	makerOrderCumulativeQuoteAssetAmountFilled: number;
-	makerOrderDirection: string;
-	makerOrderId: number;
-	makerRebate: number;
-	marketFilter: string;
-	marketIndex: number;
-	marketType: string;
-	oraclePrice: number;
-	price: number;
-	quoteAssetAmountFilled: number;
-	quoteAssetAmountSurplus: number;
-	referrerReward: number;
-	slot: number;
-	spotFulfillmentMethodFee: number;
-	symbol: string;
-	taker: string;
-	takerFee: number;
-	takerOrderBaseAssetAmount: number;
-	takerOrderCumulativeBaseAssetAmountFilled: number;
-	takerOrderCumulativeQuoteAssetAmountFilled: number;
-	takerOrderDirection: string;
-	takerOrderId: number;
-	ts: number;
-	txSig: string;
-	txSigIndex: number;
 };
 
 // Used by the subscriber client to subscribe to the candles websocket endpoint
