+++ common-ts/src/index.ts
@@ -26,8 +26,9 @@ export * from './common-ui-utils/settings/settings';
 export * from './utils/priority-fees';
 export * from './utils/orderbook';
 export * from './clients/candleClient';
-export * from './clients/tvFeed';
+export * from './clients/marketDataFeed';
 export * from './clients/swiftClient';
+export * from './clients/tvFeed';
 export * from './utils/pollingSequenceGuard';
 export * from './utils/driftEvents';
 export * from './utils/MultiplexWebSocket';
